:recycle: Long -> Short로 변경

+++ src/main/java/org/example/develociraptor/domain/position/repository/PositionJpaRepository.java
@@ -3,6 +3,6 @@ package org.example.develociraptor.domain.position.repository;
 import org.example.develociraptor.domain.position.entity.Position;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PositionJpaRepository extends JpaRepository<Position, Long> {
+public interface PositionJpaRepository extends JpaRepository<Position, Short> {
 
 }
